pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-v3"
version = "0.1.0"
description = "Advanced AI Agent System with Redis Storage and CrewAI Integration"
authors = [
    {name = "Agent V3 Team", email = "team@agent-v3.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain>=0.0.200",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "huggingface_hub>=0.15.0",
    "transformers>=4.20.0",
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "crewai>=0.1.0",
    "pydantic>=2.0.0",
    "asyncio",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/agent-v3/agent-v3"
Repository = "https://github.com/agent-v3/agent-v3"
Documentation = "https://agent-v3.readthedocs.io"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88